feat(filter-lab): browse filters as a thumbnail grid
Replaces the full-width row list with Jianying's own shape: a square
preview per filter, name and badges beneath, favourite and download as
overlays. Row-era label helpers move to jianying-filter-lab-labels so the
card component can go rather than linger unused.
